ContosoOutdoorStore/PromptFlow/.promptflow/lkg_sources/requirements.txt
azure.search.documents
azure.cosmos